requirements.txt
marker-pdf==1.7.3
pandas==2.2.3
openpyxl==3.1.5
